pyproject.toml
[tool.ruff]
line-length = 100

[tool.isort]
profile = "black"

[tool.pytest.ini_options]
filterwarnings = ["ignore::marshmallow.warnings.RemovedInMarshmallow4Warning"]

[tool.poetry]
name = "inconnu"
version = "8.4.1"
description = "An advanced Discord dice bot and character manager for V5"
authors = ["tiltowait <208040+tiltowait@users.noreply.github.com>"]
license = "MIT"

[tool.poetry.scripts]
inconnu = "main:main"

[tool.poetry.dependencies]
python = ">=3.11,<3.12"
flatdict = "^4.0.1"
motor = "^3.0.0"
numpy = "^1.23.3"
pymongo = {extras = ["srv"], version = "^4.2.0"}
python-dotenv = "^0.21.0"
umongo = "^3.1.0"
uvicorn = "^0.37.0"
validators = "^0.35.0"
jinja2 = "^3.1.2"
jinja-markdown = "^1.210911"
fastapi = "^0.119.0"
py-cord = {extras = ["speed"], version = "^2.6.1"}
pyparsing = "^3.1.1"
cachetools = "^5.5.2"
uvloop = "^0.21.0"
async-timeout = "^5.0.1"
idna = "^3.11"
loguru = "^0.7.3"

[tool.poetry.group.dev.dependencies]
pytest-asyncio = "^0.20.3"
httpx = "^0.23.3"
debugpy = "^1.6.3"
mypy = "^0.982"
pylint = "^2.15.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
